CC-682 Edited Classes lesson tests files based on the proof library changes
