+++ answers/Classes/0b0f737a-fa9c-4cf6-bbf2-dd6c4b217fde/CreateSimpleClassTest.php
