+++ tests/Classes/0b0f737a-fa9c-4cf6-bbf2-dd6c4b217fde/CreateSimpleClassTest.php
@@ -1,8 +1,16 @@
 <?php
 use CodingAvenue\Proof\Code;
-use Proofs\Proof;
-class CreateSimpleClassTest extends Proof
+use PHPUnit\Framework\TestCase;
+
+class CreateSimpleClassTest extends TestCase
 {
+	protected static $code;
+
+    public static function setupBeforeClass()
+    {
+        self::$code = new Code(getcwd() . "/" . getenv("TEST_INDEX"));
+	}
+	
 	public function testPhpStartTag()
 	{
 		$checkStart = self::$code->codeStartCheck();
