+++ tests/Classes/1f0bff5a-85ec-40cb-84d8-367b0eebb9ad/MissingFunctionKeywordTest.php
@@ -1,8 +1,16 @@
 <?php
 use CodingAvenue\Proof\Code;
-use Proofs\Proof;
-class MissingAssignmentOperatorTest extends Proof
+use PHPUnit\Framework\TestCase;
+
+class MissingFunctionKeywordTest extends TestCase
 {
+	protected static $code;
+
+    public static function setupBeforeClass()
+    {
+        self::$code = new Code(getcwd() . "/" . getenv("TEST_INDEX"));
+	}
+	
 	public function testPhpStartTag()
 	{
 		$checkStart = self::$code->codeStartCheck();
@@ -14,7 +22,7 @@ class MissingAssignmentOperatorTest extends Proof
 	{
 		$evaluator = self::$code->evaluator();
 		$evaled    = $evaluator->evaluate();
-		$expected  = "This is a class property.";
+		$expected  = "This is a class method.";
 
 		$this->assertEquals($expected, $evaled['output'], "Expected output is \"$expected\".");
 	}
@@ -23,7 +31,7 @@ class MissingAssignmentOperatorTest extends Proof
 	{
 		$nodes = self::$code->find('construct[name="echo"]');
 
-		$this->assertEquals(2, $nodes->count(), "Expecting two echo statements.");
+		$this->assertEquals(1, $nodes->count(), "Expecting a single echo statement.");
 	}
 
 	public function testAssignment()
@@ -72,10 +80,10 @@ class MissingAssignmentOperatorTest extends Proof
 		$this->assertEquals(1, $nodes->count(), "Expecting a class declaration of the `MyClass` class.");
 	}  
 
-	public function testMyPropCall()
+	public function testMyMethodCall()
 	{
-		$myProp = self::$code->find('property-call[name="myProp", variable="myObject"]');
+		$myProp = self::$code->find('method-call[name="myMethod", variable="myObject"]');
 
-		$this->assertEquals(1, $myProp->count(), "Expecting a 'myProp' property call of 'myObject'.");
+		$this->assertEquals(1, $myProp->count(), "Expecting a 'myMethod()' method call of 'myObject'.");
 	}
 }
