+++ tests/Classes/355f59d8-93c5-4399-bf42-ca1d010ee92e/MissingAssignmentOperatorTest.php
@@ -1,8 +1,16 @@
 <?php
 use CodingAvenue\Proof\Code;
-use Proofs\Proof;
-class MissingFunctionKeywordTest extends Proof
+use PHPUnit\Framework\TestCase;
+
+class MissingAssignmentOperatorTest extends TestCase
 {
+	protected static $code;
+
+    public static function setupBeforeClass()
+    {
+        self::$code = new Code(getcwd() . "/" . getenv("TEST_INDEX"));
+	}
+	
 	public function testPhpStartTag()
 	{
 		$checkStart = self::$code->codeStartCheck();
@@ -14,7 +22,7 @@ class MissingFunctionKeywordTest extends Proof
 	{
 		$evaluator = self::$code->evaluator();
 		$evaled    = $evaluator->evaluate();
-		$expected  = "This is a class method.";
+		$expected  = "This is a class property.";
 
 		$this->assertEquals($expected, $evaled['output'], "Expected output is \"$expected\".");
 	}
@@ -23,7 +31,7 @@ class MissingFunctionKeywordTest extends Proof
 	{
 		$nodes = self::$code->find('construct[name="echo"]');
 
-		$this->assertEquals(1, $nodes->count(), "Expecting a single echo statement.");
+		$this->assertEquals(2, $nodes->count(), "Expecting two echo statements.");
 	}
 
 	public function testAssignment()
@@ -72,10 +80,10 @@ class MissingFunctionKeywordTest extends Proof
 		$this->assertEquals(1, $nodes->count(), "Expecting a class declaration of the `MyClass` class.");
 	}  
 
-	public function testMyMethodCall()
+	public function testMyPropCall()
 	{
-		$myProp = self::$code->find('method-call[name="myMethod", variable="myObject"]');
+		$myProp = self::$code->find('property-call[name="myProp", variable="myObject"]');
 
-		$this->assertEquals(1, $myProp->count(), "Expecting a 'myMethod()' method call of 'myObject'.");
+		$this->assertEquals(1, $myProp->count(), "Expecting a 'myProp' property call of 'myObject'.");
 	}
 }
