+++ tests/Classes/3ce1a9ec-c263-41a9-8792-b2a21f45ec7d/CorrectMultipleErrorsTest.php
@@ -1,8 +1,16 @@
 <?php
 use CodingAvenue\Proof\Code;
-use Proofs\Proof;
-class CorrectMultipleErrorsTest extends Proof
+use PHPUnit\Framework\TestCase;
+
+class CorrectMultipleErrorsTest extends TestCase
 {
+	protected static $code;
+
+    public static function setupBeforeClass()
+    {
+        self::$code = new Code(getcwd() . "/" . getenv("TEST_INDEX"));
+    }
+
 	public function testPhpStartTag()
 	{
 		$checkStart = self::$code->codeStartCheck();
