+++ tests/Classes/f37728d4-5e19-4549-85de-8d7d3874ef4f/MissingClassKeywordTest.php
@@ -1,8 +1,16 @@
 <?php
 use CodingAvenue\Proof\Code;
-use Proofs\Proof;
-class MissingParenthesesAfterMethodTest extends Proof
+use PHPUnit\Framework\TestCase;
+
+class MissingClassKeywordTest extends TestCase
 {
+	protected static $code;
+
+    public static function setupBeforeClass()
+    {
+        self::$code = new Code(getcwd() . "/" . getenv("TEST_INDEX"));
+	}
+	
 	public function testPhpStartTag()
 	{
 		$checkStart = self::$code->codeStartCheck();
@@ -14,7 +22,7 @@ class MissingParenthesesAfterMethodTest extends Proof
 	{
 		$evaluator = self::$code->evaluator();
 		$evaled    = $evaluator->evaluate();
-		$expected  = "This is a class method.";
+		$expected  = "This is a class property.";
 
 		$this->assertEquals($expected, $evaled['output'], "Expected output is \"$expected\".");
 	}
@@ -23,7 +31,7 @@ class MissingParenthesesAfterMethodTest extends Proof
 	{
 		$nodes = self::$code->find('construct[name="echo"]');
 
-		$this->assertEquals(1, $nodes->count(), "Expecting a single echo statement.");
+		$this->assertEquals(2, $nodes->count(), "Expecting two echo statements.");
 	}
 
 	public function testAssignment()
@@ -52,7 +60,7 @@ class MissingParenthesesAfterMethodTest extends Proof
 		$obj = self::$code->find('class[name="MyClass"]');
 		$subNodes = $obj->getSubnode();
 		$myMethod = $subNodes->find('method[name="myMethod"]');
-		
+
 		$this->assertEquals(1, $myMethod->count(), "Expecting a myMethod() method.");
 	}
 
@@ -61,7 +69,7 @@ class MissingParenthesesAfterMethodTest extends Proof
 		$obj = self::$code->find('class[name="MyClass"]');
 		$subNodes = $obj->getSubnode();
 		$myProp = $subNodes->find('property[name="myProp", type="public"]');
-
+		
 		$this->assertEquals(1, $myProp->count(), "Expecting a public class property named 'myProp'.");
 	}
 
@@ -72,10 +80,10 @@ class MissingParenthesesAfterMethodTest extends Proof
 		$this->assertEquals(1, $nodes->count(), "Expecting a class declaration of the `MyClass` class.");
 	}  
 
-	public function testMyMethodCall()
+	public function testMyPropCall()
 	{
-		$myProp = self::$code->find('method-call[name="myMethod", variable="myObject"]');
+		$myProp = self::$code->find('property-call[name="myProp", variable="myObject"]');
 
-		$this->assertEquals(1, $myProp->count(), "Expecting a 'myMethod()' method call of 'myObject'.");
+		$this->assertEquals(1, $myProp->count(), "Expecting a 'myProp' property call of 'myObject'.");
 	}
 }
